Remove redux and set publicPath for webpack
Redux is not being used now. So will remove from bloat

Issues with refreshing a page that was navigated by react-router. Upon refresh a 404 error was given when trying to perform a GET request. This is because the app uses CSR (client side routing). Traditionally routing is handled by making GET requests to a server. CSR apps insteads use JS to hijack the traditional routing methods and control navigation. This only works when the app has successfully loaded. The app will not initially load if it is not on the base route. To fix this webpack config must explicitly detail the base route and the condition at which it will fall back onto it. See below:
https://ui.dev/react-router-cannot-get-url-refresh/