Add default obj & remove unnecessary fields & func
Add default objects to the item components certification, education, workexperience. This makes it clearer what fields are needed when adding a new element.

Removed unneccessary fields on education and old & unused functions in work expereince

Fixed react key warning on skill page. Problem was the key was assigned to child and not parent container in map