Set  html height to fill the space of the pdf page
For resumes which will exceed more than 1 page. The html will now fill up the remaining height. Previously on the last page th styling was cut short by the size of the HTML. Now the html height on print will increase to fill up to 4 A4 page sizes if neccessary.

Kept the print media queries in the ResumeTemplateLoader as it handles the display of the resume's html.

Because it is difficult to determine the dpi height of a device. We set a hidden element with the height set at 297mm (height of an a4 element). We then compare the resume's height with this hidden element & scale the height as neccessary