Add protected route & remove withAuthorization
Using protected route instead of HO WithAuthorization component because it is clearer which route is protected.

Needed to add the firebase listener to the app component so uth context can update.

Also needed to add an isReady parameter to ProtectedRoute component. This was needed because the re-direct in the
protectedRoute was evaluated before firebase listener could evaluate whether the use was authenticated or not. This meant on refresh of a protectedRoute it would always re-direct to sign-in. IsReady lets the protectedRoute know the listener has been evaluated.