Fix spacing when pagebreak on last child container
The last container does not have a sibling to apply the margin-top styling to. Generally what follows the last container is a new section. On print media, we will apply a margin-top to all sections, remove previous margins & paddings on containers, and add margin-bottom on the continaners. This is done because prior to print we needed more padding to fit the "Page Break" buttons. However on print we do not need the padding anymore, however we need consistent separation between containers, sections and the section when it starts on a new page.