Changed where the resume's id & renamed the field
Resume's id is set at the start of the form instead of on save. This will ensure all resumes will have an id field. This is also cleaner, so the id field doesn't need to be set after save etc.

Also renamed identifier field on resume from resumeId -> id, to be consistent with other models