Add loading on download & basic error catch on finalise
Created a preload component which will display a loading animation. This animation is called on load of the finalisation page after filling out neccessary fields. This is so the user knows something is happening in the background and is not left idle

Basic error check DOES NOT check if the resume model is valid