Remove repetitive functions on resume sections
Using the UseStateHelper functions in education, certification, personal and education sections. This cleans a lot of duplicated code across all files. Also, all section props dealing with events will be prefixed with "on" when passing to the component. During using of the prop in the component, the prop is then re-prefixed with "handle"